run_flake.py/postprocess.py: move stuff into separate file
